fix(frontend/shared): move notification message from shared to frontend tm-42
