+++ apps/frontend/src/libs/modules/notification/libs/enums/notification-message.enum.ts
@@ -1,6 +1,8 @@
 const NotificationMessage = {
+	AVATAR_CHANGED: "Your avatar has been successfully changed.",
 	FRIEND_FOLLOW_SUCCESS: "Followed successfully.",
 	FRIEND_UNFOLLOW_SUCCESS: "Unfollowed successfully.",
+	PROFILE_UPDATED: "Your profile has been updated.",
 } as const;
 
 export { NotificationMessage };
