+++ apps/frontend/src/modules/files/libs/enums/enums.ts
@@ -1,1 +1,1 @@
-export { FilesApiPath, NotificationMessage } from "shared";
+export { FilesApiPath } from "shared";
