+++ apps/frontend/src/modules/files/slices/actions.ts
@@ -1,8 +1,8 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
+import { NotificationMessage } from "~/libs/modules/notification/notification.js";
 import { type AsyncThunkConfig } from "~/libs/types/types.js";
 
-import { NotificationMessage } from "../libs/enums/enums.js";
 import { type UserFileResponseDto } from "../libs/types/type.js";
 import { name as sliceName } from "./files.slice.js";
