+++ apps/frontend/src/modules/users/libs/enums/enums.ts
@@ -1,1 +1,1 @@
-export { NotificationMessage, UsersApiPath } from "shared";
+export { UsersApiPath } from "shared";
