+++ packages/shared/src/libs/enums/enums.ts
@@ -2,5 +2,4 @@ export { APIPath } from "./api-path.enum.js";
 export { AppEnvironment } from "./app-environment.enum.js";
 export { ContentType } from "./content-type.enum.js";
 export { ExceptionMessage } from "./exception-message.enum.js";
-export { NotificationMessage } from "./notification-message.enum.js";
 export { ServerErrorType } from "./server-error-type.enum.js";
