--- packages/shared/src/libs/enums/notification-message.enum.ts
@@ -1,6 +1,0 @@
-const NotificationMessage = {
-	AVATAR_CHANGED: "Your avatar has been successfully changed.",
-	PROFILE_UPDATED: "Your profile has been updated.",
-} as const;
-
-export { NotificationMessage };
